refactor(App.vue): rename QRCodeCapture component to QRCodeScan
