+++ src/components/QRCodeScan.vue
